Make tests for Navbar. This was an, eh, adventure?
