+++ frontend/package.json
@@ -33,6 +33,7 @@
     "@react-router/dev": "^7.12.0",
     "@tailwindcss/vite": "^4.1.13",
     "@testing-library/dom": "^10.4.1",
+    "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.2",
     "@types/node": "^22",
     "@types/react": "^19.2.14",
