--- frontend/tests/globalSetup.ts
@@ -1,21 +1,0 @@
-import { beforeEach, vi } from "vitest";
-
-export function setup() {}
-
-export function teardown() {}
-
-export function setupLocalStorage() {
-  const store: Record<string, string> = {};
-  vi.stubGlobal("localStorage", {
-    getItem: (key: string) => store[key] ?? null,
-    setItem: (key: string, value: string) => {
-      store[key] = value;
-    },
-    clear: () => {
-      Object.keys(store).forEach((k) => delete store[k]);
-    },
-  });
-  beforeEach(() => {
-    localStorage.clear();
-  });
-}
