+++ frontend/vitest.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
   test: {
     environment: "jsdom",
     reporters: ["default", "hanging-process"],
-    globalSetup: ["./tests/globalSetup.ts"],
+	setupFiles: ["/tests/setup.ts"],
     env: {
       NODE_OPTIONS: "--no-experimental-webstorage",
     },
